Clang: Replace macro usage with HostOsInfo

+++ src/libs/clangbackendipc/connectionclient.cpp
@@ -30,6 +30,8 @@
 #include "cmbregistertranslationunitsforeditormessage.h"
 #include "cmbunregistertranslationunitsforeditormessage.h"
 
+#include <utils/hostosinfo.h>
+
 #include <QCoreApplication>
 #include <QMetaMethod>
 #include <QProcess>
@@ -180,13 +182,10 @@ void ConnectionClient::endProcess(QProcess *process)
 
 void ConnectionClient::terminateProcess(QProcess *process)
 {
-    Q_UNUSED(process)
-#ifndef Q_OS_WIN32
-    if (isProcessIsRunning()) {
+    if (!Utils::HostOsInfo::isWindowsHost() && isProcessIsRunning()) {
         process->terminate();
         process->waitForFinished();
     }
-#endif
 }
 
 void ConnectionClient::killProcess(QProcess *process)
